requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
groq
redis